separate network banner
