+++ apps/sv/frontend/src/components/layout/NetworkBanner.tsx
@@ -1,0 +1,32 @@
+import * as React from 'react';
+
+import { Stack, Typography } from '@mui/material';
+
+import { useNetworkInstanceName } from '../../hooks';
+
+const NetworkBanner: React.FC = () => {
+  const networkInstanceName = useNetworkInstanceName();
+  const knownColors = ['mainnet', 'testnet', 'devnet', 'scratchnet', 'localnet'];
+  const networkInstanceNameColor = knownColors.includes(networkInstanceName.toLowerCase())
+    ? `colors.${networkInstanceName.toLowerCase()}`
+    : 'colors.neutral.30';
+  return (
+    <Stack
+      direction="row"
+      alignItems="center"
+      justifyContent="center"
+      sx={{
+        backgroundColor: networkInstanceNameColor,
+        color: 'black',
+        height: '50px',
+        width: '100%',
+      }}
+    >
+      <Typography id="network-instance-name" data-testid="network-instance-name" variant="h6">
+        <b>You are on {networkInstanceName} </b>
+      </Typography>
+    </Stack>
+  );
+};
+
+export default NetworkBanner;
