+++ apps/sv/frontend/src/components/layout/SvNavigationShell.tsx
@@ -2,10 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 import * as React from 'react';
 
-import { Box, Stack, Typography } from '@mui/material';
+import { Box } from '@mui/material';
 
 import { HEADER_PB, layoutTokens, PAGE_PX } from '../../theme/tokens';
-import { useNetworkInstanceName } from '../../hooks';
 import SvTopNav from './SvTopNav';
 import { SvNavLinkItem } from './SvNavLink';
 
@@ -20,12 +19,6 @@ interface SvNavigationShellProps {
  * Dev Mode: padding-bottom 64px, background #272727.
  */
 const SvNavigationShell: React.FC<SvNavigationShellProps> = ({ navLinks, onLogout, pageName }) => {
-  const networkInstanceName = useNetworkInstanceName();
-  const knownColors = ['mainnet', 'testnet', 'devnet', 'scratchnet', 'localnet'];
-  const networkInstanceNameColor = knownColors.includes(networkInstanceName.toLowerCase())
-    ? `colors.${networkInstanceName.toLowerCase()}`
-    : 'colors.neutral.30';
-
   return (
     <Box
       data-component="navigation"
@@ -38,25 +31,6 @@ const SvNavigationShell: React.FC<SvNavigationShellProps> = ({ navLinks, onLogou
         width: '100%',
       }}
     >
-      <Stack
-        direction="row"
-        alignItems="center"
-        justifyContent="center"
-        sx={{
-          position: 'sticky',
-          top: 0,
-          zIndex: 1100,
-          pointerEvents: 'none',
-          backgroundColor: networkInstanceNameColor,
-          color: 'black',
-          height: '50px',
-          width: '100%',
-        }}
-      >
-        <Typography id="network-instance-name" data-testid="network-instance-name" variant="h6">
-          <b>You are on {networkInstanceName} </b>
-        </Typography>
-      </Stack>
       <Box sx={{ px: PAGE_PX, width: '100%' }}>
         <SvTopNav navLinks={navLinks} onLogout={onLogout} />
       </Box>
